Move ProcessChecker to tribler-common
