+++ src/tribler-core/run_tribler_headless.py
@@ -12,9 +12,9 @@ from datetime import date
 from socket import inet_aton
 
 from tribler_common.osutils import get_appstate_dir
+from tribler_common.process_checker import ProcessChecker
 
 from tribler_core.config.tribler_config import TriblerConfig
-from tribler_core.modules.process_checker import ProcessChecker
 from tribler_core.start_core import Session
 from tribler_core.utilities.path_util import Path
