+++ src/tribler-core/tribler_core/utilities/tiny_tribler_service.py
@@ -5,8 +5,9 @@ from pathlib import Path
 
 from ipv8.loader import IPv8CommunityLoader
 
+from tribler_common.process_checker import ProcessChecker
+
 from tribler_core.config.tribler_config import TriblerConfig
-from tribler_core.modules.process_checker import ProcessChecker
 from tribler_core.start_core import Session
